pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "recipes"
version = "0.0.1"
description = "Recipe Website"
authors = [ "Kyle King <https://github.com/KyleKing>",]
repository = "https://github.com/kyleking/recipes"
license = "MIT"

[tool.isort]
balanced_wrapping = true
default_section = "THIRDPARTY"
force_grid_wrap = 0
length_sort = false
line_length = 120

[tool.poetry.dependencies]
python = "^3.8,<3.9"
mkdocs = "*"
mkdocs-git-revision-date-plugin = "*"
mkdocs-material = "*"
optimize-images = "*"
pillow = "*"

[tool.poetry.dependencies.calcipy]
git = "https://github.com/KyleKing/calcipy.git"
branch = "dev/development"

[tool.poetry.dev-dependencies.calcipy]
git = "https://github.com/KyleKing/calcipy.git"
branch = "dev/development"
extras = [ "development",]
# [tool.poetry.dev-dependencies]
# calcipy = {version = "*", extras = [ "development",]}
